// frontend/src/lib/api/api.interfaces.d.ts
export interface IGetUserResponse {
    email: string;
    full_name: string;
    disabled: boolean;
    auth_method: 'email' | 'google';
    photo: string;
    id: string;
    has_api_key: boolean;
}
// ...
export interface Member {
    id: string;
    organization_id: string;
    role: OrganizationRoleEnum;
    user: Omit<IGetUserResponse, 'auth_method'>;
    created_at: Date;
    updated_at: Date | null;
}
export interface AddMemberPayload {
    user_id: string;
    organization_id: string;
    role: OrganizationRoleEnum;
}
export interface UpdateMemberPayload {
    role: OrganizationRoleEnum;
}
export interface CreateInvitePayload {
    email: string;
    role: OrganizationRoleEnum;
    organization_id: string;
}
export interface Orbit {
    id: string;
    name: string;
    organization_id: string;
    total_members: number;
    created_at: Date;
    updated_at: Date | null;
    bucket_secret_id: string;
    total_collections: number;
    role: OrbitRoleEnum;
    permissions: OrbitPermissions;
    total_artifacts: number;
}
export interface CreateOrbitPayload {
    name: string;
    bucket_secret_id: string;
    members: {
        user_id: string;
        role: OrbitRoleEnum;
    }[];
    notify: boolean;
}
export interface UpdateOrbitPayload {
    id: string;
    name: string;
    bucket_secret_id: string;
}
export interface AddMemberToOrbitPayload {
    user_id: string;
    orbit_id: string;
    role: OrbitRoleEnum;
}
export interface OrbitDetails extends Orbit {
    members: OrbitMember[];
}
export interface OrbitMember extends Omit<Member, 'organization_id' | 'role'> {
    orbit_id: string;
    role: OrbitRoleEnum;
}
// ...
export interface OrbitPermissions {
    orbit: Omit<PermissionEnum, PermissionEnum.deploy>;
    orbit_user: Omit<PermissionEnum, PermissionEnum.deploy>;
    artifact: PermissionEnum;
    collection: PermissionEnum;
    track: Omit<PermissionEnum, PermissionEnum.deploy>;
}
export declare enum PermissionEnum {
    list = "list",
    read = "read",
    create = "create",
    update = "update",
    delete = "delete",
    deploy = "deploy",
    leave = "leave"
}